ml_api/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
numpy>=1.23.5
tensorflow-cpu>=2.12.0
pillow>=9.5.0
python-dotenv>=1.0.0
python-jose==3.3.0
passlib==1.7.4
requests==2.26.0
geopy==2.2.0 